[ECS] Rename --level option to --set
