+++ packages/EasyCodingStandard/src/Console/Command/CheckCommand.php
@@ -102,7 +102,7 @@ final class CheckCommand extends Command
         if ($totalCheckersLoaded === 0) {
             throw new NoCheckersLoadedException(
                 'No checkers were found. Register them in your config in "services:" '
-                . 'section, load them via "--config <file>.yml" or "--level <level> option.'
+                . 'section, load them via "--config <file>.yml" or "--set <set>" option.'
             );
         }
     }
